+++ src/state.rs
@@ -39,8 +39,8 @@ impl std::fmt::Display for StatePersistenceError {
 
 impl State {
     pub fn update_from_context(self, me: ExtendedMe) -> Self {
-        let ws_id = self
-            .default_workspace
+        let ws_id = me
+            .default_workspace_id
             .filter(|&ws| me.workspaces.iter().any(|w| w.id == ws))
             .or_else(|| me.workspaces.first().map(|ws| ws.id));
         let project_id = self
@@ -183,7 +183,10 @@ impl State {
 
     pub async fn save_customization(&self) -> NetResult<()> {
         let client = Client::from_api_token(&self.api_token);
-        self.customization.clone().save(&client).await
+        self.customization
+            .clone()
+            .save(self.default_workspace, &client)
+            .await
     }
 }
 
@@ -194,6 +197,7 @@ mod test {
     use super::State;
     use crate::customization::WeekDay;
     use crate::entities::{Preferences, Workspace, WorkspaceId};
+    use crate::test::test_client;
     use crate::time_entry::TimeEntry;
     use crate::ExtendedMe;
 
@@ -220,11 +224,11 @@ mod test {
             ..TimeEntry::default()
         };
         let me = ExtendedMe {
-            api_token: "token".to_string(),
             projects: vec![],
             workspaces: vec![ws.clone()],
             time_entries: vec![e_running.clone(), e_stopped, e_foreign.clone()],
             beginning_of_week: 0,
+            default_workspace_id: Some(WorkspaceId::new(1)),
             preferences: Preferences::default(),
         };
         let mut state = State::default().update_from_context(me);
@@ -260,11 +264,11 @@ mod test {
             ..TimeEntry::default()
         };
         let me = ExtendedMe {
-            api_token: "token".to_string(),
             projects: vec![],
             workspaces: vec![ws.clone()],
             time_entries: vec![e.clone()],
             beginning_of_week: 0,
+            default_workspace_id: Some(WorkspaceId::new(1)),
             preferences: Preferences::default(),
         };
         let state = State::default().update_from_context(me);
@@ -279,11 +283,11 @@ mod test {
     #[test]
     fn test_state_preferences() {
         let me = ExtendedMe {
-            api_token: "token".to_string(),
             projects: vec![],
             workspaces: vec![],
             time_entries: vec![],
             beginning_of_week: 2, // Tue
+            default_workspace_id: Some(WorkspaceId::new(1)),
             preferences: Preferences::default(),
         };
         let state = State::default().update_from_context(me);
@@ -291,5 +295,56 @@ mod test {
             state.customization.week_start_day,
             WeekDay(chrono::Weekday::Tue)
         );
+    }
+
+    #[test]
+    fn test_state_workspace_default() {
+        let ws1 = Workspace {
+            id: WorkspaceId::new(10),
+            ..Workspace::default()
+        };
+        let ws2 = Workspace {
+            id: WorkspaceId::new(11),
+            ..Workspace::default()
+        };
+        let mut me = ExtendedMe {
+            projects: vec![],
+            workspaces: vec![ws1.clone(), ws2.clone()],
+            time_entries: vec![],
+            beginning_of_week: 2, // Tue
+            default_workspace_id: Some(ws2.id),
+            preferences: Preferences::default(),
+        };
+
+        let state = State::default().update_from_context(me.clone());
+        assert_eq!(state.default_workspace, Some(ws2.id));
+
+        me.default_workspace_id = None;
+        let state = State::default().update_from_context(me.clone());
+        assert_eq!(state.default_workspace, Some(ws1.id));
+
+        me.default_workspace_id = Some(WorkspaceId::new(0)); // Not found
+        let state = State::default().update_from_context(me);
+        assert_eq!(state.default_workspace, Some(ws1.id));
+    }
+
+    #[async_std::test]
+    async fn test_crud() {
+        let client = test_client();
+        let me = ExtendedMe::load(&client).await.expect("get me");
+        let state = State::default().update_from_context(me);
+        assert!(state.default_workspace.is_some());
+
+        let mut customization = state.customization;
+        let new_day = WeekDay((*customization.week_start_day).succ());
+        customization.week_start_day = new_day;
+        customization
+            .save(state.default_workspace, &client)
+            .await
+            .expect("save customization");
+
+        let me = ExtendedMe::load(&client).await.expect("get me");
+        let state = State::default().update_from_context(me);
+        assert_eq!(state.customization.week_start_day, new_day);
     }
 }