+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 	  entry after editing
 * feat: added project badge to running entry row
 * feat: added Cargo and NPM as distribution channels
+* feat: added self-update support
 
 ## 0.1.2