+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.0-beta.1
+
+Release fix
+
 ## 0.2.0-beta.0
 
 * bugfix: fixed crash when no projects exist